% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Netflix Explorer Shiny App}
\usage{
run_app()
}
\description{
This function launches the interactive Shiny app that allows users to
explore the Netflix data included with the package.
}
